requirements.txt
streamlit==1.11.1
streamlit-chat==0.0.2.1
streamlit-option-menu==0.3.1
tokenizers==0.13.3
transformers==4.30.2
virtualenv==20.16.2
pickle4==0.0.1
Pillow==9.1.1
requests==2.28.1
scikit-learn==1.3.0
pandas==1.4.3
numpy==1.22.4